refactor: 모든 전략class를 domain strategy 패키지로 이동
