+++ src/main/java/step3/domain/strategy/MovableStrategy.java
@@ -1,4 +1,4 @@
-package step3.domain;
+package step3.domain.strategy;
 
 public interface MovableStrategy {
     boolean isMove();
