+++ src/main/java/step3/domain/strategy/SelectFarthestDistanceWinnerStrategy.java
@@ -1,4 +1,6 @@
-package step3.domain;
+package step3.domain.strategy;
+
+import step3.domain.RacingCar;
 
 import java.util.List;
 import java.util.stream.Collectors;
