+++ src/test/java/step3/domain/RacingCarTest.java
@@ -3,6 +3,7 @@ package step3.domain;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
+import step3.domain.strategy.MovableStrategy;
 
 import static org.assertj.core.api.Assertions.assertThat;
